+++ decitala/database.py
@@ -32,7 +32,8 @@ from .utils import (
 	successive_ratio_array,
 	filter_single_anga_class_fragments,
 	filter_sub_fragments,
-	pitch_content_to_contour
+	pitch_content_to_contour,
+	contour_to_prime_contour
 )
 from .trees import (
 	FragmentTree,
@@ -187,18 +188,21 @@ def create_database(
 		logging.info("Connected to database at: {}".format(db_path))
 
 		cur = conn.cursor()
-		fragment_table_string = "CREATE TABLE Fragments (Onset_Start REAL, Onset_Stop REAL, Fragment BLOB, Mod TEXT, Factor REAL, Pitch_Content BLOB, Pitch_Contour BLOB, Is_Slurred INT)"
+		fragment_table_string = "CREATE TABLE Fragments (Onset_Start REAL, Onset_Stop REAL, Fragment BLOB, Mod TEXT, Factor REAL, Pitch_Content BLOB, Pitch_Contour BLOB, Prime_Contour BLOB, Is_Slurred INT)"
 		cur.execute(fragment_table_string)
 
 		for this_partition in partitioned_data:
 			for this_fragment in this_partition:
-				fragment_insertion_string = "INSERT INTO Fragments VALUES({0}, {1}, '{2}', '{3}', {4}, '{5}', '{6}', {7})".format(this_fragment["onset_range"][0], # start offset
+				contour = list(pitch_content_to_contour(this_fragment["pitch_content"]))
+				prime_contour = list(contour_to_prime_contour(contour, include_depth=False))
+				fragment_insertion_string = "INSERT INTO Fragments VALUES({0}, {1}, '{2}', '{3}', {4}, '{5}', '{6}', '{7}' {8})".format(this_fragment["onset_range"][0], # start offset
 																													this_fragment["onset_range"][1], # end offset
 																													this_fragment["fragment"], # fragment
 																													this_fragment["mod"][0], # mod type 
 																													this_fragment["mod"][1], # mod factor/difference
 																													this_fragment["pitch_content"], # pitch content
-																													list(pitch_content_to_contour(this_fragment["pitch_content"])), # pitch contour 
+																													contour, # pitch contour 
+																													prime_contour, # prime contour
 																													int(this_fragment["is_spanned_by_slur"])) # is_slurred
 				cur.execute(fragment_insertion_string)