+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 
 ## [v0.4.0] - January 2, 2020 (Kent, CT)
 ### Added
-- Contour and Prime Contour (using Morris' 1993 algorithm) implemented; now are columns in the table.
+- Added ``utils.pitch_content_to_contour``. 
+- Prime Contour (using Morris' 1993 algorithm) implemented; now are columns in the table.
 
 ## [v0.3.2] - January 1, 2020 (Kent, CT)
 ### Added